+++ integration-test/src/examples/complex-demo.vue
@@ -1,10 +1,11 @@
 <script setup lang="ts">
 import { ref } from 'vue'
-import { FormModel, JsonSchemaNode } from '@form-renderer/engine'
+import type { FormModel, JsonSchemaNode } from '@form-renderer/engine'
 import type { ComponentDefinition } from '@form-renderer/adapter-vue3'
-import { FormRenderer } from '@form-renderer/starter-element-plus'
 import { commonFieldEventMapping } from '@form-renderer/preset-element-plus'
-
+import { FormRenderer } from '@form-renderer/starter-element-plus'
+import { FormEngine } from '@form-renderer/engine'
+console.log('FormEngine', FormEngine)
 import DynamicField from '@/custom/DynamicField.vue'
 
 import { ComplexDemoSchema, ComplexDemoModel } from '@/cases'
@@ -12,7 +13,6 @@ const formRef = ref()
 
 const formData = ref<FormModel>(ComplexDemoModel)
 const formSchema = ref<JsonSchemaNode>(ComplexDemoSchema)
-
 const extraComponents: ComponentDefinition[] = [
   {
     type: 'field',