+++ packages/AdapterVue3/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @form-renderer/adapter-vue3
 
+## 1.0.0-alpha.5
+
+### Patch Changes
+
+- package.json调整
+- Updated dependencies
+  - @form-renderer/engine@1.0.0-alpha.5
+
 ## 1.0.0-alpha.0
 
 ### Major Changes