+++ packages/Engine/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @form-renderer/engine
 
+## 1.0.0-alpha.5
+
+### Patch Changes
+
+- package.json调整
+
 ## 1.0.0-alpha.0
 
 ### Major Changes